+++ modules/apps/CardHandsList.mjs
@@ -153,6 +153,7 @@ export class CardHandsList extends HandlebarsApplicationMixin(ApplicationV2) {
         // Drop a Card
         for (const cards of this.element.querySelectorAll(`.cards`)) {
             cards.addEventListener('drop', CardHandsList.#onDropCard.bind(this));
+            cards.addEventListener('contextmenu', CardHandsList.#onFlipCard.bind(this));
         }
 
         if (!foundry.utils.isEmpty(this.scrollLeftPositions)) {
@@ -237,11 +238,11 @@ export class CardHandsList extends HandlebarsApplicationMixin(ApplicationV2) {
     static async #onOpenCardActions(event, target) {
         // Prevent multiple executions
         event.preventDefault();
-        const card = fromUuidSync(target.dataset.uuid);
+        const document = fromUuidSync(target.dataset.uuid);
 
-        if (card) {
+        if (document) {
             new CardActionsSheet({
-                document: card,
+                document,
                 buttonActions: CONFIG.CardHandsList.menuItems.cardContextOptions,
             }).render(true);
         }
@@ -251,11 +252,11 @@ export class CardHandsList extends HandlebarsApplicationMixin(ApplicationV2) {
     static async #onOpenHand(event, target) {
         // Prevent multiple executions
         event.preventDefault();
-        const hand = game.cards.get(target.dataset.id);
+        const document = game.cards.get(target.dataset.id);
 
-        if (hand) {
+        if (document) {
             new HandActionsSheet({
-                document: hand,
+                document,
                 buttonActions: CONFIG.CardHandsList.menuItems.handContextOptions,
             }).render({force: true});
         }
@@ -322,7 +323,7 @@ export class CardHandsList extends HandlebarsApplicationMixin(ApplicationV2) {
         const defaultMode = hand.getFlag(handsModule.id, 'default-draw-mode');
         const faceDown = hand.getFlag(handsModule.id, 'face-down');
 
-        if (defaultDeck || defaultMode) {
+        if (defaultDeck) {
             const deck = game.cards.get(defaultDeck);
             const cardsInHand = hand.cards.contents;
             const sort = cardsInHand.length ? cardsInHand.reverse()[0].sort + 10 : 0;
@@ -339,6 +340,11 @@ export class CardHandsList extends HandlebarsApplicationMixin(ApplicationV2) {
             await hand.drawDialog();
         }
     };
+    // Flip a Card in a Cards Hand
+    static async #onFlipCard(event) {
+        const card = await fromUuid(event.target.dataset.uuid);
+        await card.flip();
+    }
 
     // Drag a Card in a Cards Hand
     static async #onDragCard(event) {